chore(file): revert fileId to Long type

+++ src/main/scala/com/secretapp/backend/data/message/rpc/file/FileLocation.scala
@@ -3,7 +3,7 @@ package com.secretapp.backend.data.message.rpc.file
 import com.secretapp.backend.data.message.ProtobufMessage
 import com.reactive.messenger.{ api => protobuf }
 
-case class FileLocation(fileId: Int, accessHash: Long) extends ProtobufMessage
+case class FileLocation(fileId: Long, accessHash: Long) extends ProtobufMessage
 {
   def toProto = protobuf.FileLocation(fileId, accessHash)
 }
